+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Yarn now accepts sponsorships! Please give a look at our [OpenCollective](https:
 
 - The PnP linker now schedules packages to be rebuilt if their unplugged folder is removed
 - Plugins can now access `yup` again to make migration easier - will be removed again in the future
+- `yarn config unset` will now correctly unset non-nested properties
 
 ### Shell